use double instead of int for some properties of Country object

+++ Milestone Project 13 To 15/MilestoneP13To15/Data Layer/Country.swift
@@ -17,9 +17,9 @@ class Country: Codable {
     var altSpellings: [String]?
     var subregion, region: String?
     var population: Int?
-    var latlng: [Int]?
+    var latlng: [Double]?
     var demonym: String?
-    var area: Int?
+    var area: Double?
     var timezones, borders: [String]?
     var nativeName, numericCode: String?
     var flag: String?
